Allow renaming a buffer to a permanent new name
Using setzerOS/exwm-rename-buffer, you can permanently set a buffer's
name for an EXWM buffer that won't change even if the X application
decides to change the window title. This combines well with
find-or-create, so you can have multiple named instances of applications
(e.g. running terminal sessions) that won't be found by it.